formatting and linux system roots test platform fix
- system roots test for linux now only runs on linux systems
