+++ test/core/security/linux_system_roots_test.cc
@@ -24,10 +24,10 @@
 #include <grpc/support/alloc.h>
 #include <grpc/support/log.h>
 #include <grpc/support/string_util.h>
-#include <string.h>
 #include <sys/param.h>
 
 #include "src/core/lib/gpr/env.h"
+#include "src/core/lib/gpr/string.h"
 #include "src/core/lib/gpr/tmpfile.h"
 #include "src/core/lib/iomgr/load_file.h"
 #include "src/core/lib/security/context/security_context.h"
@@ -39,62 +39,55 @@
 #include "src/core/tsi/transport_security.h"
 #include "test/core/util/test_config.h"
 
-#include "gtest/gtest.h"
-
 #ifndef GRPC_USE_SYSTEM_SSL_ROOTS_ENV_VAR
 #define GRPC_USE_SYSTEM_SSL_ROOTS_ENV_VAR "GRPC_USE_SYSTEM_SSL_ROOTS"
 #endif
 
-namespace grpc {
-namespace {
-
-TEST(AbsoluteFilePathTest, ConcatenatesCorrectly) {
+// Test GetAbsoluteFilePath.
+static void test_absolute_cert_path() {
   const char* directory = "nonexistent/test/directory";
   const char* filename = "doesnotexist.txt";
-  char result_path[MAXPATHLEN];
-  grpc_core::GetAbsoluteFilePath(directory, filename, result_path);
-  EXPECT_STREQ(result_path, "nonexistent/test/directory/doesnotexist.txt");
+  const char* result_path = static_cast<char*>(gpr_malloc(MAXPATHLEN));
+  grpc_core::GetAbsoluteFilePath(result_path, directory, filename);
+  GPR_ASSERT(
+      strcmp(result_path, "nonexistent/test/directory/doesnotexist.txt") == 0);
+  gpr_free((char*)result_path);
 }
 
-TEST(CreateRootCertsBundleTest, ReturnsEmpty) {
-  // Test that CreateRootCertsBundle returns an empty slice for null or
-  // nonexistent cert directories.
-  grpc_slice result_slice = grpc_core::CreateRootCertsBundle(nullptr);
-  EXPECT_TRUE(GRPC_SLICE_IS_EMPTY(result_slice));
-  grpc_slice_unref(result_slice);
-  result_slice = grpc_core::CreateRootCertsBundle("does/not/exist");
-  EXPECT_TRUE(GRPC_SLICE_IS_EMPTY(result_slice));
-  grpc_slice_unref(result_slice);
-}
-
-TEST(CreateRootCertsBundleTest, BundlesCorrectly) {
+static void test_cert_bundle_creation() {
   gpr_setenv(GRPC_USE_SYSTEM_SSL_ROOTS_ENV_VAR, "true");
-  // Test that CreateRootCertsBundle returns a correct slice.
+
+  /* Test that CreateRootCertsBundle returns a correct slice. */
   grpc_slice roots_bundle = grpc_empty_slice();
   GRPC_LOG_IF_ERROR(
-      "load_file",
-      grpc_load_file("test/core/security/etc/bundle.pem", 1, &roots_bundle));
-  // result_slice should have the same content as roots_bundle.
+      "load_file", grpc_load_file("test/core/security/etc/bundle/bundle.pem", 1,
+                                  &roots_bundle));
+  /* result_slice should have the same content as roots_bundle. */
   grpc_slice result_slice =
-      grpc_core::CreateRootCertsBundle("test/core/security/etc/test_roots");
+      grpc_core::CreateRootCertsBundle("test/core/security/etc/roots");
   char* result_str = grpc_slice_to_c_string(result_slice);
   char* bundle_str = grpc_slice_to_c_string(roots_bundle);
-  EXPECT_STREQ(result_str, bundle_str);
-  // Clean up.
+  GPR_ASSERT(strcmp(result_str, bundle_str) == 0);
+  // TODO: fix bazel builds by introducing working absolute path
+  if (GRPC_SLICE_IS_EMPTY(result_slice)) {
+    printf("Your build is unsupported by 'test_cert_bundle_creation'\n");
+  }
+  /* TODO: add tests for branches in CreateRootCertsBundle that return empty
+   * slices. */
+
+  /* Cleanup. */
   unsetenv(GRPC_USE_SYSTEM_SSL_ROOTS_ENV_VAR);
+  unsetenv("GRPC_SYSTEM_SSL_ROOTS_DIR");
   gpr_free(result_str);
   gpr_free(bundle_str);
   grpc_slice_unref(roots_bundle);
   grpc_slice_unref(result_slice);
 }
 
-}  // namespace
-}  // namespace grpc
-
-int main(int argc, char** argv) {
-  grpc_test_init(argc, argv);
-  ::testing::InitGoogleTest(&argc, argv);
-  return RUN_ALL_TESTS();
+int main() {
+  test_absolute_cert_path();
+  test_cert_bundle_creation();
+  return 0;
 }
 #else
 int main() {
